+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## v1.1.0 (2024-01-05)
+
+- Rename get_token to kc_get_token
+- Add auth_scope option
+
+
 ## v1.0.0 (2024-01-04)
 
 - Init release with get_token module